vision-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
pytesseract==0.3.13
Pillow==10.3.0
google-cloud-vision==3.7.2
pydantic==2.7.1
httpx==0.27.0
